:recycle: Refactor : move RunningConst to common package
